add euclidean distance process for next iteration

+++ perhitungan/views.py
@@ -5,6 +5,7 @@ from django.core import serializers
 from django.shortcuts import redirect, render
 from perhitungan.models import Perhitungan
 from rm.models import Penyakit, RM
+import numpy as np
 
 # Create your views here.
 
@@ -12,7 +13,7 @@ from rm.models import Penyakit, RM
 def index(request):
     if not request.user.is_authenticated:
         return redirect('login')
-    perhitungan = Perhitungan.objects.all()
+    perhitungan = Perhitungan.objects.filter(iteration=1)
     if request.is_ajax():
         perhitungan_serialized = serializers.serialize('python', perhitungan)
         perhitungan_json = [d['fields'] for d in perhitungan_serialized]
@@ -21,6 +22,8 @@ def index(request):
 
 
 def generate(request):
+    if not request.user.is_authenticated:
+        return redirect('login')
     Perhitungan.objects.all().delete()
     penyakit = Penyakit.objects.values_list('nama_penyakit', flat=True)
     for p in penyakit:
@@ -59,6 +62,8 @@ def generate(request):
 
 
 def proses(request):
+    if not request.user.is_authenticated:
+        return redirect('login')
     dataSet = Perhitungan.objects.filter(iteration=1).values_list()
     dataCount = Perhitungan.objects.filter(iteration=1).count()
     initCentroid = Perhitungan.objects.order_by('?')[:3].values_list()
@@ -118,5 +123,114 @@ def proses(request):
     for d, data in zip(range(dataCount), dataSet):
         Perhitungan.objects.filter(
             iteration=1, id=data[0]).update(kelompok=kelompok[d])
+
+    # Iteration Count
+    dataSet = Perhitungan.objects.filter(iteration=1).values_list()
+    i = 2
+    while not Perhitungan.objects.filter(iteration__gt=1) or i != 3:
+        maxIteration = Perhitungan.objects.all().values_list(
+            'iteration', flat=True).order_by('-iteration')[0]
+
+        for data in dataSet:
+            Perhitungan(
+                nama_penyakit=data[1], jumlah_perempuan=data[2], jumlah_laki=data[3], usia_anak=data[4], usia_remaja=data[5], usia_dewasa=data[6], usia_tua=data[7], usia_manula=data[8], iteration=i).save()
+
+        # --> Get cluster 1 centroid
+        centroid1 = Perhitungan.objects.filter(
+            kelompok='C1', iteration=i-1).values_list()
+        centroid1_list = list()
+        for c in range(2, 9):
+            column_temp = 0
+            for data in centroid1:
+                print(data[c], data[1])
+                column_temp += data[c]
+            centroid1_list.append(column_temp)
+        print(centroid1_list)
+
+        # --> Get cluster 2 centroid
+        centroid2 = Perhitungan.objects.filter(
+            kelompok='C2', iteration=i-1).values_list()
+        centroid2_list = list()
+        for c in range(2, 9):
+            column_temp = 0
+            for data in centroid2:
+                print(data[c], data[1])
+                column_temp += data[c]
+            centroid2_list.append(column_temp)
+        print(centroid2_list)
+
+        # --> Get cluster 3 centroid
+        centroid3 = Perhitungan.objects.filter(
+            kelompok='C3', iteration=i-1).values_list()
+        centroid3_list = list()
+        for c in range(2, 9):
+            column_temp = 0
+            for data in centroid3:
+                print(data[c], data[1])
+                column_temp += data[c]
+            centroid3_list.append(column_temp)
+        print(centroid3_list)
+
+        # --> Euclidean Distance for iteration > 1
+        dataSet = Perhitungan.objects.filter(iteration=i).values_list()
+        dataCount = Perhitungan.objects.filter(iteration=i).count()
+
+        # --> Cluster 1
+        centroid1 = centroid1_list
+        cluster1 = list()
+        for n in range(dataCount):
+            cluster1_list = 0
+            cluster1_temp = 0
+            for l, r in zip(range(2, 9), range(7)):
+                cluster1_list = math.pow(dataSet[n][l] - centroid1[r], 2)
+                cluster1_temp += cluster1_list
+            cluster1.append(math.sqrt(cluster1_temp))
+        for d, data in zip(range(dataCount), dataSet):
+            Perhitungan.objects.filter(
+                iteration=i, id=data[0]).update(c1=cluster1[d])
+
+        # --> Cluster 2
+        centroid2 = centroid2_list
+        cluster2 = list()
+        for n in range(dataCount):
+            cluster2_list = 0
+            cluster2_temp = 0
+            for l, r in zip(range(2, 9), range(7)):
+                cluster2_list = math.pow(dataSet[n][l] - centroid2[r], 2)
+                cluster2_temp += cluster2_list
+            cluster2.append(math.sqrt(cluster2_temp))
+        for d, data in zip(range(dataCount), dataSet):
+            Perhitungan.objects.filter(
+                iteration=i, id=data[0]).update(c2=cluster2[d])
+
+        # --> Cluster 3
+        centroid3 = centroid3_list
+        cluster3 = list()
+        for n in range(dataCount):
+            cluster3_list = 0
+            cluster3_temp = 0
+            for l, r in zip(range(2, 9), range(7)):
+                cluster3_list = math.pow(dataSet[n][l] - centroid3[r], 2)
+                cluster3_temp += cluster3_list
+            cluster3.append(math.sqrt(cluster3_temp))
+        for d, data in zip(range(dataCount), dataSet):
+            Perhitungan.objects.filter(
+                iteration=i, id=data[0]).update(c3=cluster3[d])
+
+        # --> Get Cluster Group
+        dataSet = Perhitungan.objects.filter(iteration=i).values_list()
+        kelompok = list()
+        for d in range(dataCount):
+            if (float(dataSet[d][9]) <= float(dataSet[d][10])) and (float(dataSet[d][9]) <= float(dataSet[d][11])):
+                kelompok.append('C1')
+            elif (float(dataSet[d][10]) <= float(dataSet[d][9])) and (float(dataSet[d][10]) <= float(dataSet[d][11])):
+                kelompok.append('C2')
+            elif (float(dataSet[d][11]) <= float(dataSet[d][9])) and (float(dataSet[d][11]) <= float(dataSet[d][10])):
+                kelompok.append('C3')
+        for d, data in zip(range(dataCount), dataSet):
+            Perhitungan.objects.filter(
+                iteration=i, id=data[0]).update(kelompok=kelompok[d])
+
+        i += 1
 
     return HttpResponse(dataSet)
